test(config): add missing error path tests for GPU and GGUF detection
- Consolidated tests/test_config_coverage.py into tests/test_config.py
- Added comprehensive tests for _detect_gpu and _has_gguf_support including error paths
- Achieved 99% coverage for src/mnemo_mcp/config.py
- Followed 1:1 mapping rule for source and test modules
