+++ tests/test_config.py
@@ -1,9 +1,11 @@
 """Tests for mnemo_mcp.config — Settings, API keys, embedding resolution."""
 
 import os
+import sys
 from pathlib import Path
+from unittest.mock import MagicMock, patch
 
-from mnemo_mcp.config import Settings
+from mnemo_mcp.config import Settings, _detect_gpu, _has_gguf_support
 
 
 class TestSettingsDefaults:
@@ -282,3 +284,164 @@ class TestGoogleDriveClientId:
         )
         s = Settings(api_keys=None)
         assert s.google_drive_client_id == "123456.apps.googleusercontent.com"
+
+class TestGpuDetection:
+    def test_detect_gpu_success_cuda(self):
+        mock_ort = MagicMock()
+        mock_ort.get_available_providers.return_value = ["CUDAExecutionProvider"]
+        with patch.dict(sys.modules, {"onnxruntime": mock_ort}):
+            assert _detect_gpu() is True
+
+    def test_detect_gpu_success_dml(self):
+        mock_ort = MagicMock()
+        mock_ort.get_available_providers.return_value = ["DmlExecutionProvider"]
+        with patch.dict(sys.modules, {"onnxruntime": mock_ort}):
+            assert _detect_gpu() is True
+
+    def test_detect_gpu_no_gpu_providers(self):
+        mock_ort = MagicMock()
+        mock_ort.get_available_providers.return_value = ["CPUExecutionProvider"]
+        with patch.dict(sys.modules, {"onnxruntime": mock_ort}):
+            assert _detect_gpu() is False
+
+    def test_detect_gpu_import_error(self):
+        """_detect_gpu returns False when onnxruntime is not installed."""
+        with patch.dict(sys.modules, {"onnxruntime": None}):
+            assert _detect_gpu() is False
+
+    def test_detect_gpu_runtime_exception(self):
+        """_detect_gpu returns False when get_available_providers raises."""
+        mock_ort = MagicMock()
+        mock_ort.get_available_providers.side_effect = Exception("Runtime error")
+        with patch.dict(sys.modules, {"onnxruntime": mock_ort}):
+            assert _detect_gpu() is False
+
+
+class TestGgufSupport:
+    def test_has_gguf_support_success(self):
+        with patch.dict(sys.modules, {"llama_cpp": MagicMock()}):
+            assert _has_gguf_support() is True
+
+    def test_has_gguf_support_import_error(self):
+        with patch.dict(sys.modules, {"llama_cpp": None}):
+            assert _has_gguf_support() is False
+
+
+class TestResolveEmbeddingBackendAuto:
+    def test_auto_detect_with_api_keys(self):
+        """Auto-detect returns 'cloud' when API keys are set."""
+        s = Settings(
+            api_keys="COHERE_API_KEY:test",
+            embedding_backend="",
+        )
+        assert s.resolve_embedding_backend() == "cloud"
+
+    def test_auto_detect_with_no_config(self, monkeypatch):
+        """Auto-detect returns 'local' when nothing is configured."""
+        for k in (
+            "JINA_AI_API_KEY",
+            "GEMINI_API_KEY",
+            "GOOGLE_API_KEY",
+            "OPENAI_API_KEY",
+            "COHERE_API_KEY",
+            "CO_API_KEY",
+            "XAI_API_KEY",
+        ):
+            monkeypatch.delenv(k, raising=False)
+        s = Settings(embedding_backend="", api_keys=None)
+        assert s.resolve_embedding_backend() == "local"
+
+    def test_litellm_backward_compat(self):
+        """'litellm' embedding_backend maps to 'cloud'."""
+        s = Settings(embedding_backend="litellm", api_keys=None)
+        assert s.resolve_embedding_backend() == "cloud"
+
+    def test_litellm_rerank_backend_backward_compat(self):
+        """'litellm' rerank_backend maps to 'cloud'."""
+        s = Settings(rerank_backend="litellm", api_keys=None)
+        assert s.resolve_rerank_backend() == "cloud"
+
+
+class TestResolveLocalModelVariants:
+    def test_resolve_local_embedding_model_variants(self):
+        with (
+            patch("mnemo_mcp.config._detect_gpu", return_value=True),
+            patch("mnemo_mcp.config._has_gguf_support", return_value=True),
+        ):
+            s = Settings(api_keys=None)
+            assert "GGUF" in s.resolve_local_embedding_model()
+
+        with (
+            patch("mnemo_mcp.config._detect_gpu", return_value=False),
+            patch("mnemo_mcp.config._has_gguf_support", return_value=True),
+        ):
+            s = Settings(api_keys=None)
+            assert "ONNX" in s.resolve_local_embedding_model()
+
+        with (
+            patch("mnemo_mcp.config._detect_gpu", return_value=False),
+            patch("mnemo_mcp.config._has_gguf_support", return_value=False),
+        ):
+            s = Settings(api_keys=None)
+            assert "ONNX" in s.resolve_local_embedding_model()
+
+    def test_resolve_local_rerank_model_variants(self):
+        with (
+            patch("mnemo_mcp.config._detect_gpu", return_value=True),
+            patch("mnemo_mcp.config._has_gguf_support", return_value=True),
+        ):
+            s = Settings(api_keys=None)
+            assert "GGUF" in s.resolve_local_rerank_model()
+
+        with (
+            patch("mnemo_mcp.config._detect_gpu", return_value=False),
+            patch("mnemo_mcp.config._has_gguf_support", return_value=False),
+        ):
+            s = Settings(api_keys=None)
+            assert "ONNX" in s.resolve_local_rerank_model()
+
+class TestConfigCoverageExtra:
+    def test_setup_providers_sdk_mode(self, monkeypatch):
+        s = Settings(api_keys="GOOGLE_API_KEY:test-key")
+        mode = s.setup_providers()
+        assert mode == "sdk"
+        monkeypatch.delenv("GOOGLE_API_KEY", raising=False)
+        monkeypatch.delenv("GEMINI_API_KEY", raising=False)
+
+    def test_setup_providers_local_mode(self, monkeypatch):
+        for k in (
+            "JINA_AI_API_KEY",
+            "GEMINI_API_KEY",
+            "GOOGLE_API_KEY",
+            "OPENAI_API_KEY",
+            "COHERE_API_KEY",
+            "CO_API_KEY",
+            "XAI_API_KEY",
+        ):
+            monkeypatch.delenv(k, raising=False)
+        s = Settings(api_keys=None)
+        mode = s.setup_providers()
+        assert mode == "local"
+
+    def test_resolve_rerank_backend_custom(self):
+        s = Settings(rerank_backend="custom", api_keys=None)
+        assert s.resolve_rerank_backend() == "custom"
+
+    def test_resolve_rerank_backend_cloud_via_api_keys(self):
+        s = Settings(api_keys="COHERE_API_KEY:test")
+        assert s.resolve_rerank_backend() == "cloud"
+
+    def test_resolve_rerank_model_via_api_keys(self):
+        s = Settings(api_keys="COHERE_API_KEY:test")
+        assert s.resolve_rerank_model() == "rerank-v4.0-pro"
+
+    def test_resolve_provider_mode_env_vars(self, monkeypatch):
+        """Test resolve_provider_mode by setting one of the env vars."""
+        for k in ["JINA_AI_API_KEY", "GEMINI_API_KEY", "GOOGLE_API_KEY", "OPENAI_API_KEY", "COHERE_API_KEY", "CO_API_KEY", "XAI_API_KEY"]:
+            monkeypatch.delenv(k, raising=False)
+
+        s = Settings(api_keys=None)
+        assert s.resolve_provider_mode() == "local"
+
+        monkeypatch.setenv("XAI_API_KEY", "test")
+        assert s.resolve_provider_mode() == "sdk"
